Modularize main and cleanup data collector
